Extract attribute interfaces into a dedicated type definition file
